group functional blocks in subfolders
