+++ SampleApplications/SDK/Opc.Ua.Client.ComplexTypes/TypeBuilder/AssemblyModule.cs
