+++ SampleApplications/SDK/Opc.Ua.Client.ComplexTypes/TypeBuilder/AttributeExtensions.cs
