+++ SampleApplications/SDK/Opc.Ua.Client.ComplexTypes/TypeBuilder/ComplexTypeBuilder.cs
@@ -36,45 +36,6 @@ using System.Reflection.Emit;
 namespace Opc.Ua.Client.ComplexTypes
 {
     /// <summary>
-    /// Factory function for the default complex type builder
-    /// using classes created with Reflection.Emit.
-    /// </summary>
-    public class ComplexTypeBuilderFactory : 
-        IComplexTypeBuilderFactory
-    {
-        AssemblyModule m_moduleFactory;
-
-        /// <summary>
-        /// Factory creates types in the assembly module.
-        /// </summary>
-        public ComplexTypeBuilderFactory(string assemblyName = null)
-        {
-            m_moduleFactory = new AssemblyModule(assemblyName);
-        }
-
-        /// <summary>
-        /// Create a new type builder which uses Reflection.Emit.
-        /// </summary>
-        public override IComplexTypeBuilder Create(
-            string targetNamespace,
-            int targetNamespaceIndex,
-            string moduleName = null)
-        {
-            return new ComplexTypeBuilder(
-                m_moduleFactory, targetNamespace,
-                targetNamespaceIndex, moduleName);
-        }
-
-        /// <summary>
-        /// Return array of all types created in this factory.
-        /// </summary>
-        public override Type[] GetTypes()
-        {
-            return m_moduleFactory.GetTypes();
-        }
-    }
-
-    /// <summary>
     /// Build an assembly with custom enum types and 
     /// complex types based on the BaseComplexType class
     /// using System.Reflection.Emit.
@@ -264,98 +225,4 @@ namespace Opc.Ua.Client.ComplexTypes
         #endregion
     }
 
-    /// <summary>
-    /// Helper to build property fields.
-    /// </summary>
-    public class ComplexTypeFieldBuilder : IComplexTypeFieldBuilder
-    {
-        #region Constructors
-        public ComplexTypeFieldBuilder(
-            TypeBuilder structureBuilder,
-            StructureType structureType)
-        {
-            m_structureBuilder = structureBuilder;
-            m_structureType = structureType;
-        }
-        #endregion
-
-        #region Public Properties
-        /// <summary>
-        /// Build the StructureTypeId attribute for a complex type.
-        /// </summary>
-        public void AddTypeIdAttribute(
-            ExpandedNodeId complexTypeId,
-            ExpandedNodeId binaryEncodingId,
-            ExpandedNodeId xmlEncodingId
-            )
-        {
-            m_structureBuilder.StructureTypeIdAttribute(
-                complexTypeId,
-                binaryEncodingId,
-                xmlEncodingId
-                );
-        }
-
-        /// <summary>
-        /// Create a property field of a class with get and set.
-        /// </summary>
-        public void AddField(StructureField field, Type fieldType, int order)
-        {
-            var fieldBuilder = m_structureBuilder.DefineField("_" + field.Name, fieldType, FieldAttributes.Private);
-            var propertyBuilder = m_structureBuilder.DefineProperty(
-                field.Name,
-                PropertyAttributes.None,
-                fieldType,
-                null);
-            var methodAttributes =
-                System.Reflection.MethodAttributes.Public |
-                System.Reflection.MethodAttributes.HideBySig |
-                System.Reflection.MethodAttributes.Virtual;
-
-            var setBuilder = m_structureBuilder.DefineMethod("set_" + field.Name, methodAttributes, null, new[] { fieldType });
-            var setIl = setBuilder.GetILGenerator();
-            setIl.Emit(OpCodes.Ldarg_0);
-            setIl.Emit(OpCodes.Ldarg_1);
-            setIl.Emit(OpCodes.Stfld, fieldBuilder);
-            if (m_structureType == StructureType.Union)
-            {
-                // set the union selector to the new field index
-                FieldInfo unionField = typeof(UnionComplexType).GetField(
-                    "m_switchField",
-                    BindingFlags.NonPublic |
-                    BindingFlags.Instance);
-                setIl.Emit(OpCodes.Ldarg_0);
-                setIl.Emit(OpCodes.Ldc_I4, order);
-                setIl.Emit(OpCodes.Stfld, unionField);
-            }
-            setIl.Emit(OpCodes.Ret);
-
-            var getBuilder = m_structureBuilder.DefineMethod("get_" + field.Name, methodAttributes, fieldType, Type.EmptyTypes);
-            var getIl = getBuilder.GetILGenerator();
-            getIl.Emit(OpCodes.Ldarg_0);
-            getIl.Emit(OpCodes.Ldfld, fieldBuilder);
-            getIl.Emit(OpCodes.Ret);
-
-            propertyBuilder.SetGetMethod(getBuilder);
-            propertyBuilder.SetSetMethod(setBuilder);
-            propertyBuilder.DataMemberAttribute(field.Name, false, order);
-            propertyBuilder.StructureFieldAttribute(field);
-        }
-
-        /// <summary>
-        /// Finish the type creation and returns the new type.
-        /// </summary>
-        public Type CreateType()
-        {
-            var complexType = m_structureBuilder.CreateType();
-            m_structureBuilder = null;
-            return complexType;
-        }
-        #endregion
-
-        #region Private Member
-        private TypeBuilder m_structureBuilder;
-        private StructureType m_structureType;
-        #endregion
-    }
 }//namespace
