+++ SampleApplications/SDK/Opc.Ua.Client.ComplexTypes/Types/OptionalFieldsComplexType.cs
@@ -29,40 +29,36 @@
 
 
 using System;
+using System.Collections.Generic;
 using System.Linq;
-using System.Reflection;
 using System.Text;
 
 namespace Opc.Ua.Client.ComplexTypes
 {
-    public class UnionComplexType : BaseComplexType
+    public class OptionalFieldsComplexType : BaseComplexType
     {
         #region Constructors
         /// <summary>
         /// Initializes the object with default values.
         /// </summary>
-        public UnionComplexType() : base()
+        public OptionalFieldsComplexType()
         {
-            m_switchField = 0;
+            m_encodingMask = 0;
         }
 
         /// <summary>
         /// Initializes the object with a <paramref name="typeId"/>.
         /// </summary>
         /// <param name="typeId">The type to copy and create an instance from</param>
-        public UnionComplexType(ExpandedNodeId typeId) : base(typeId)
+        public OptionalFieldsComplexType(ExpandedNodeId typeId) : base(typeId)
         {
-            m_switchField = 0;
+            m_encodingMask = 0;
         }
         #endregion
 
         #region Public Properties
-        /// <summary>
-        /// The union selector determines which property is valid.
-        /// A value of 0 means all properties are invalid, x=1..n means the
-        /// xth property is valid.
-        /// </summary>
-        UInt32 SwitchField => m_switchField;
+
+        UInt32 EncodingMask => m_encodingMask;
 
         /// <summary>
         /// Makes a deep copy of the object.
@@ -72,35 +68,35 @@ namespace Opc.Ua.Client.ComplexTypes
         /// </returns>
         public override object MemberwiseClone()
         {
-            UnionComplexType clone = (UnionComplexType)base.MemberwiseClone();
-            clone.m_switchField = m_switchField;
+            OptionalFieldsComplexType clone = (OptionalFieldsComplexType)base.MemberwiseClone();
+            clone.m_encodingMask = m_encodingMask;
             return clone;
         }
 
         /// <summary cref="IEncodeable.Encode(IEncoder)" />
         public override void Encode(IEncoder encoder)
         {
+            bool useReversibleEncoding = GetUseReversibleEncoding(encoder);
+
             encoder.PushNamespace(TypeId.NamespaceUri);
 
-            encoder.WriteUInt32("SwitchField", m_switchField);
-            if (m_switchField != 0)
+            if (useReversibleEncoding)
             {
-                int unionSelector = 1;
-                int valueRank = -1;
-                PropertyInfo unionProperty = null;
-                foreach (var property in GetPropertyEnumerator())
-                {
-                    if (unionSelector == m_switchField)
-                    {
-                        valueRank = property.ValueRank;
-                        unionProperty = property.PropertyInfo;
-                        break;
-                    }
-                    unionSelector++;
-                }
-                EncodeProperty(encoder, unionProperty, valueRank);
+                encoder.WriteUInt32("EncodingMask", m_encodingMask);
             }
 
+            foreach (var property in GetPropertyEnumerator())
+            {
+                if (property.IsOptional)
+                {
+                    if ((property.OptionalFieldMask & m_encodingMask) == 0)
+                    {
+                        continue;
+                    }
+                }
+
+                EncodeProperty(encoder, property.PropertyInfo, property.ValueRank);
+            }
             encoder.PopNamespace();
         }
 
@@ -109,19 +105,19 @@ namespace Opc.Ua.Client.ComplexTypes
         {
             decoder.PushNamespace(TypeId.NamespaceUri);
 
-            m_switchField = decoder.ReadUInt32("SwitchField");
+            m_encodingMask = decoder.ReadUInt32("EncodingMask");
 
-            UInt32 unionSelector = m_switchField;
-            if (unionSelector > 0)
+            foreach (var property in GetPropertyEnumerator())
             {
-                foreach (var property in GetPropertyEnumerator())
+                if (property.IsOptional)
                 {
-                    if (--unionSelector == 0)
+                    if ((property.OptionalFieldMask & m_encodingMask) == 0)
                     {
-                        DecodeProperty(decoder, property.PropertyInfo, property.ValueRank);
-                        break;
+                        continue;
                     }
                 }
+
+                DecodeProperty(decoder, property.PropertyInfo, property.ValueRank);
             }
             decoder.PopNamespace();
         }
@@ -134,13 +130,13 @@ namespace Opc.Ua.Client.ComplexTypes
                 return true;
             }
 
-            var valueBaseType = equalValue as UnionComplexType;
+            var valueBaseType = equalValue as OptionalFieldsComplexType;
             if (valueBaseType == null)
             {
                 return false;
             }
 
-            if (SwitchField != valueBaseType.SwitchField)
+            if (m_encodingMask != valueBaseType.EncodingMask)
             {
                 return false;
             }
@@ -151,21 +147,22 @@ namespace Opc.Ua.Client.ComplexTypes
                 return false;
             }
 
-            if (m_switchField != 0)
+            foreach (var property in GetPropertyEnumerator())
             {
-                UInt32 unionSelector = m_switchField;
-                foreach (var property in GetPropertyEnumerator())
+                if (property.IsOptional)
                 {
-                    if (--unionSelector == 0)
+                    if ((property.OptionalFieldMask & m_encodingMask) == 0)
                     {
-                        if (!Utils.IsEqual(property.GetValue(this), property.GetValue(valueBaseType)))
-                        {
-                            return false;
-                        }
-                        break;
+                        continue;
                     }
                 }
+
+                if (!Utils.IsEqual(property.GetValue(this), property.GetValue(valueBaseType)))
+                {
+                    return false;
+                }
             }
+
             return true;
         }
         #endregion
@@ -185,18 +182,17 @@ namespace Opc.Ua.Client.ComplexTypes
             if (format == null)
             {
                 StringBuilder body = new StringBuilder();
-                if (m_switchField != 0)
+                foreach (var property in GetPropertyEnumerator())
                 {
-                    UInt32 unionSelector = m_switchField;
-                    foreach (var property in GetPropertyEnumerator())
+                    if (property.IsOptional)
                     {
-                        if (--unionSelector == 0)
+                        if ((property.OptionalFieldMask & m_encodingMask) == 0)
                         {
-                            object unionProperty = property.GetValue(this);
-                            AppendPropertyValue(formatProvider, body, unionProperty, property.ValueRank);
-                            break;
+                            continue;
                         }
                     }
+
+                    AppendPropertyValue(formatProvider, body, property.GetValue(this), property.ValueRank);
                 }
 
                 if (body.Length > 0)
@@ -221,79 +217,54 @@ namespace Opc.Ua.Client.ComplexTypes
         /// <summary>
         /// Access property values by index.
         /// </summary>
-        /// <remarks>
-        /// The value of a Union is determined by the union selector.
-        /// Calling get on an unselected property returns null, 
-        ///     otherwise the selected object.
-        /// Calling get with an invalid index (e.g.-1) returns the selected object.
-        /// Calling set with a valid object on a selected property sets the value and the 
-        /// union selector.
-        /// Calling set with a null object or an invalid index unselects the union.
-        /// </remarks>
         public override object this[int index]
         {
             get
             {
-                if (index + 1 == (int)m_switchField)
+                var property = m_propertyList.ElementAt(index);
+                if (property.IsOptional &&
+                    (property.OptionalFieldMask & m_encodingMask) == 0)
                 {
-                    return m_propertyList.ElementAt(index).GetValue(this);
+                    return null;
                 }
-                if (index < 0 &&
-                    m_switchField > 0)
-                {
-                    return m_propertyList.ElementAt((int)m_switchField - 1).GetValue(this);
-                }
-                return null;
+                return property.GetValue(this);
             }
             set
             {
-                if (index >= 0)
+                var property = m_propertyList.ElementAt(index);
+                property.SetValue(this, value);
+                if (property.IsOptional)
                 {
-                    m_propertyList.ElementAt(index).SetValue(this, value);
-                    // note: selector is updated in SetValue by emitted code for union
-                    // m_unionSelector = (uint)(index + 1);
-                    if (value != null)
+                    if (value == null)
                     {
-                        return;
+                        m_encodingMask &= ~property.OptionalFieldMask;
                     }
-                    // reset union selector if value is a null
+                    else
+                    {
+                        m_encodingMask |= property.OptionalFieldMask;
+                    }
                 }
-                m_switchField = 0;
             }
         }
 
         /// <summary>
         /// Access property values by name.
         /// </summary>
-        /// <remarks>
-        /// The value of a Union is determined by the union selector.
-        /// Calling get on an unselected property returns null, 
-        /// otherwise the selected object.
-        /// Calling get with an invalid name returns the selected object.
-        /// Calling set with a valid object on a selected property sets the value and the 
-        /// union selector. 
-        /// Calling set with a null object or an invalid name unselects the union.
-        /// </remarks>
         public override object this[string name]
         {
             get
             {
-                if (SwitchField > 0)
+                ComplexTypePropertyAttribute property;
+                if (m_propertyDict.TryGetValue(name, out property))
                 {
-                    ComplexTypePropertyAttribute property;
-                    if (m_propertyDict.TryGetValue(name, out property))
+                    if (property.IsOptional &&
+                        (property.OptionalFieldMask & m_encodingMask) == 0)
                     {
-                        if ((int)m_switchField == property.Order)
-                        {
-                            return property.GetValue(this);
-                        }
+                        return null;
                     }
-                    else
-                    {
-                        return m_propertyList.ElementAt((int)SwitchField - 1).GetValue(this);
-                    }
+                    return property.GetValue(this);
                 }
-                return null;
+                throw new KeyNotFoundException();
             }
             set
             {
@@ -301,21 +272,44 @@ namespace Opc.Ua.Client.ComplexTypes
                 if (m_propertyDict.TryGetValue(name, out property))
                 {
                     property.SetValue(this, value);
-                    // note: selector is updated in SetValue by emitted code for union
-                    // m_unionSelector = (uint)(property.Order);
-                    if (value != null)
+                    if (value == null)
                     {
-                        return;
+                        m_encodingMask &= ~property.OptionalFieldMask;
                     }
-                    // reset union selector if value is a null
+                    else
+                    {
+                        m_encodingMask |= property.OptionalFieldMask;
+                    }
                 }
-                m_switchField = 0;
+                else
+                {
+                    throw new KeyNotFoundException();
+                }
+            }
+        }
+        #endregion
+
+        #region Private Members
+        protected override void InitializePropertyAttributes()
+        {
+            base.InitializePropertyAttributes();
+
+            // build optional field mask attribute
+            UInt32 optionalFieldMask = 1;
+            foreach (var property in GetPropertyEnumerator())
+            {
+                property.OptionalFieldMask = 0;
+                if (property.IsOptional)
+                {
+                    property.OptionalFieldMask = optionalFieldMask;
+                    optionalFieldMask <<= 1;
+                }
             }
         }
         #endregion
 
         #region Private Fields
-        protected UInt32 m_switchField;
+        protected UInt32 m_encodingMask;
         #endregion
     }
