+++ SampleApplications/SDK/Opc.Ua.Client.ComplexTypes/Types/StructureDefinitionAttribute.cs
