+++ SampleApplications/SDK/Opc.Ua.Client.ComplexTypes/Types/StructureFieldAttribute.cs
