+++ SampleApplications/SDK/Opc.Ua.Client.ComplexTypes/Types/StructureTypeAttribute.cs
